frontend/requirements.txt
streamlit==1.9.0
google.auth==2.6.6